// src/app/core/models/famille-produit.ts
export class FamilleProduit {
    id_famille: string | undefined;
    codificationNationale: string | undefined;
    description: string| undefined;
    codificationMetier: string| undefined;
    date_debut: string | null| undefined;
    date_fin: string | null| undefined;
    auditUser: string| undefined;
}